+++ sdk/eventhubs/azure-messaging-eventhubs/src/test/java/com/azure/messaging/eventhubs/EventHubConsumerClientTest.java
@@ -69,6 +69,7 @@ public class EventHubConsumerClientTest {
     private static final String EVENT_HUB_NAME = "event-hub-name";
     private static final String CONSUMER_GROUP = "consumer-group-test";
     private static final String PARTITION_ID = "partition-id";
+    private static final String CLIENT_ID = "my-client-id";
     private static final ExecutorService EXECUTOR_SERVICE = Executors.newFixedThreadPool(4);
     private static final Duration TIMEOUT = Duration.ofSeconds(30);
 
@@ -123,7 +124,7 @@ public class EventHubConsumerClientTest {
         connectionStates.next(AmqpEndpointState.ACTIVE);
 
         when(connection.createReceiveLink(any(), argThat(name -> name.endsWith(PARTITION_ID)),
-            any(EventPosition.class), any(ReceiveOptions.class))).thenReturn(
+            any(EventPosition.class), any(ReceiveOptions.class), CLIENT_ID)).thenReturn(
             Mono.fromCallable(() -> {
                 System.out.println("Returning first link");
                 return amqpReceiveLink;
@@ -135,7 +136,7 @@ public class EventHubConsumerClientTest {
         when(connection.closeAsync()).thenReturn(Mono.empty());
 
         asyncConsumer = new EventHubConsumerAsyncClient(HOSTNAME, EVENT_HUB_NAME,
-            connectionProcessor, messageSerializer, CONSUMER_GROUP, PREFETCH, false, onClientClosed);
+            connectionProcessor, messageSerializer, CONSUMER_GROUP, PREFETCH, false, onClientClosed, CLIENT_ID);
         consumer = new EventHubConsumerClient(asyncConsumer, Duration.ofSeconds(10));
     }
 
@@ -161,7 +162,7 @@ public class EventHubConsumerClientTest {
         // Arrange
         final EventHubConsumerAsyncClient runtimeConsumer = new EventHubConsumerAsyncClient(
             HOSTNAME, EVENT_HUB_NAME, connectionProcessor, messageSerializer, CONSUMER_GROUP,
-            PREFETCH, false, onClientClosed);
+            PREFETCH, false, onClientClosed, CLIENT_ID);
         final EventHubConsumerClient consumer = new EventHubConsumerClient(runtimeConsumer, Duration.ofSeconds(5));
         final int numberOfEvents = 10;
         sendMessages(messageProcessor, numberOfEvents, PARTITION_ID);
@@ -190,7 +191,8 @@ public class EventHubConsumerClientTest {
         // Arrange
         final ReceiveOptions options = new ReceiveOptions().setTrackLastEnqueuedEventProperties(true);
         final EventHubConsumerAsyncClient runtimeConsumer = new EventHubConsumerAsyncClient(
-            HOSTNAME, EVENT_HUB_NAME, connectionProcessor, messageSerializer, CONSUMER_GROUP, PREFETCH, false, onClientClosed);
+            HOSTNAME, EVENT_HUB_NAME, connectionProcessor, messageSerializer, CONSUMER_GROUP, PREFETCH,
+            false, onClientClosed, CLIENT_ID);
         final EventHubConsumerClient consumer = new EventHubConsumerClient(runtimeConsumer, Duration.ofSeconds(5));
 
         final int numberOfEvents = 10;