+++ sdk/eventhubs/azure-messaging-eventhubs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.12.0-beta.1 (Unreleased)
 
 ### Features Added
+- Added identifier to client. ([#22981](https://github.com/Azure/azure-sdk-for-java/issues/22981))
 
 ### Breaking Changes